+++ lib/shopify/src/shopify_custom.dart
@@ -7,7 +7,6 @@ class ShopifyCustom with ShopifyError {
   ShopifyCustom._();
 
   GraphQLClient? get _graphQLClient => ShopifyConfig.graphQLClient;
-  GraphQLClient? get _graphQLClientAdmin => ShopifyConfig.graphQLClientAdmin;
 
   /// Singleton instance of [ShopifyCustom]
   static final ShopifyCustom instance = ShopifyCustom._();
@@ -16,20 +15,16 @@ class ShopifyCustom with ShopifyError {
   ///
   /// Returns the data of the custom query.
   ///
-  /// [adminAccess] is optional, if set to true, the admin access token will be used.
   Future<Map<String, dynamic>?> customQuery({
     required String gqlQuery,
     Map<String, dynamic> variables = const {},
-    bool adminAccess = false,
   }) async {
     final QueryOptions _options = WatchQueryOptions(
       document: gql(gqlQuery),
       variables: variables,
       fetchPolicy: ShopifyConfig.fetchPolicy,
     );
-    final QueryResult result = adminAccess
-        ? await _graphQLClientAdmin!.query(_options)
-        : await _graphQLClient!.query(_options);
+    final QueryResult result = await _graphQLClient!.query(_options);
     checkForError(result);
     return result.data;
   }
@@ -38,19 +33,15 @@ class ShopifyCustom with ShopifyError {
   ///
   /// Returns the data of the custom mutation.
   ///
-  /// [adminAccess] is optional, if set to true, the admin access token will be used.
   Future<Map<String, dynamic>?> customMutation({
     required String gqlMutation,
     Map<String, dynamic> variables = const {},
-    bool adminAccess = false,
   }) async {
     final MutationOptions _options = MutationOptions(
       document: gql(gqlMutation),
       variables: variables,
     );
-    final QueryResult result = adminAccess
-        ? await _graphQLClientAdmin!.mutate(_options)
-        : await _graphQLClient!.mutate(_options);
+    final QueryResult result = await _graphQLClient!.mutate(_options);
     checkForError(result);
     return result.data;
   }