+++ lib/shopify_flutter.dart
@@ -2,5 +2,6 @@ library;
 
 export 'shopify/shopify.dart';
 export 'enums/enums.dart';
+export 'mixins/mixins.dart';
 export 'models/models.dart';
 export 'shopify_config.dart';